bricata_api_client/tests/test_bricata_api_client.py
#!/usr/bin/env python3.8
"""Base API Client: Test Request Debug
Copyright © 2019 Jerod Gawne <https://github.com/jerodg/>

This program is free software: you can redistribute it and/or modify
it under the terms of the Server Side Public License (SSPL) as
published by MongoDB, Inc., either version 1 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
SSPL for more details.

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

You should have received a copy of the SSPL along with this program.
If not, see <https://www.mongodb.com/licensing/server-side-public-license>."""
import time

import pytest
from os import getenv

from base_api_client.base_api_utils import bprint
from bricata_api_client.bricata_api_client import BricataApiClient


@pytest.mark.asyncio
async def test_login():
    ts = time.perf_counter()
    bprint('Test: Login')

    with BricataApiClient(cfg=f'{getenv("HOME")}/.config/bricata_api_client.toml') as bac:
        results = await bac.login()
        # print('results:\n', results)  # debug

        assert type(results) is dict
        assert results['success'] is not None
        assert results['failure'][0] is None

    print('Success Result:')
    print(*results['success'], sep='\n')
    print('\nFailure Result:')
    print(*results['failure'], sep='\n')

    bprint(f'-> Completed in {(time.perf_counter() - ts):f} seconds.')
